@@ -32,10 +32,6 @@ public:
         _image.loadFile("Nyan.png");
         _sprite.setImage(_image);
         _sprite.setScale(2);
-        _prog.attach(lm::Shader("fragment_ex02.glsl", lm::Shader::Type::Fragment));
-        _prog.attach(lm::Shader("vertex_ex02.glsl", lm::Shader::Type::Vertex));
-        _prog.link();
-        _prog.use();
     }
 
     void
@@ -60,7 +56,7 @@ public:
     }
 
     void
-    render() const
+    render()
     {
         lm::SpriteBatch sb;
         
