@@ -40,7 +40,7 @@
 #include <Lums/Image.hpp>
 #include <Lums/Key.hpp>
 #include <Lums/OperatingSystem.hpp>
-#include <Lums/Rect2.hpp>
+#include <Lums/Rect.hpp>
 #include <Lums/Script.hpp>
 #include <Lums/Shader.hpp>
 #include <Lums/ShaderProgram.hpp>
@@ -48,8 +48,7 @@
 #include <Lums/SoundManager.hpp>
 #include <Lums/Sprite.hpp>
 #include <Lums/SpriteBatch.hpp>
-#include <Lums/Vector2.hpp>
-#include <Lums/Vector3.hpp>
+#include <Lums/Vector.hpp>
 #include <Lums/VertexArray.hpp>
 #include <Lums/Window.hpp>
 
